// migrations/snapshots/5b0cc1c16eed34e53feafa5dac988e737c440139604f50e1c489cb1679e63db8/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Event: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly creatorId: CodecTypes['pg/uuid@1']['output'];
      readonly capacity: CodecTypes['pg/int4@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly startsAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly location: CodecTypes['pg/text@1']['output'];
      readonly city: CodecTypes['pg/text@1']['output'];
      readonly state:
        | 'AC'
        | 'AL'
        | 'AP'
        | 'AM'
        | 'BA'
        | 'CE'
        | 'DF'
        | 'ES'
        | 'GO'
        | 'MA'
        | 'MT'
        | 'MS'
        | 'MG'
        | 'PA'
        | 'PB'
        | 'PR'
        | 'PE'
        | 'PI'
        | 'RJ'
        | 'RN'
        | 'RS'
        | 'RO'
        | 'RR'
        | 'SC'
        | 'SP'
        | 'SE'
        | 'TO';
      readonly status: 'draft' | 'published' | 'finished' | 'canceled';
      readonly minAge: CodecTypes['pg/int4@1']['output'];
      readonly ticketPriceCents: CodecTypes['pg/int4@1']['output'];
      readonly soldTickets: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly Order: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
      readonly eventId: CodecTypes['pg/uuid@1']['output'];
      readonly paymentId: CodecTypes['pg/text@1']['output'] | null;
      readonly status: 'pending' | 'paid' | 'expired' | 'canceled';
      readonly quantity: CodecTypes['pg/int4@1']['output'];
      readonly unitPriceCents: CodecTypes['pg/int4@1']['output'];
      readonly expiresAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly Ticket: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly orderId: CodecTypes['pg/uuid@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
      readonly eventId: CodecTypes['pg/uuid@1']['output'];
      readonly uniqueCode: CodecTypes['pg/uuid@1']['output'];
      readonly checkedInAt: CodecTypes['pg/timestamptz-string@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly city: CodecTypes['pg/text@1']['output'];
      readonly state:
        | 'AC'
        | 'AL'
        | 'AP'
        | 'AM'
        | 'BA'
        | 'CE'
        | 'DF'
        | 'ES'
        | 'GO'
        | 'MA'
        | 'MT'
        | 'MS'
        | 'MG'
        | 'PA'
        | 'PB'
        | 'PR'
        | 'PE'
        | 'PI'
        | 'RJ'
        | 'RN'
        | 'RS'
        | 'RO'
        | 'RR'
        | 'SC'
        | 'SP'
        | 'SE'
        | 'TO';
      readonly cpf: CodecTypes['pg/text@1']['output'];
      readonly tier: 'free' | 'verified' | 'pro';
      readonly passwordHash: CodecTypes['pg/text@1']['output'];
      readonly birthDate: CodecTypes['pg/date-temporal@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Event: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly creatorId: CodecTypes['pg/uuid@1']['input'];
      readonly capacity: CodecTypes['pg/int4@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly startsAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly location: CodecTypes['pg/text@1']['input'];
      readonly city: CodecTypes['pg/text@1']['input'];
      readonly state:
        | 'AC'
        | 'AL'
        | 'AP'
        | 'AM'
        | 'BA'
        | 'CE'
        | 'DF'
        | 'ES'
        | 'GO'
        | 'MA'
        | 'MT'
        | 'MS'
        | 'MG'
        | 'PA'
        | 'PB'
        | 'PR'
        | 'PE'
        | 'PI'
        | 'RJ'
        | 'RN'
        | 'RS'
        | 'RO'
        | 'RR'
        | 'SC'
        | 'SP'
        | 'SE'
        | 'TO';
      readonly status: 'draft' | 'published' | 'finished' | 'canceled';
      readonly minAge: CodecTypes['pg/int4@1']['input'];
      readonly ticketPriceCents: CodecTypes['pg/int4@1']['input'];
      readonly soldTickets: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly Order: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
      readonly eventId: CodecTypes['pg/uuid@1']['input'];
      readonly paymentId: CodecTypes['pg/text@1']['input'] | null;
      readonly status: 'pending' | 'paid' | 'expired' | 'canceled';
      readonly quantity: CodecTypes['pg/int4@1']['input'];
      readonly unitPriceCents: CodecTypes['pg/int4@1']['input'];
      readonly expiresAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly Ticket: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly orderId: CodecTypes['pg/uuid@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
      readonly eventId: CodecTypes['pg/uuid@1']['input'];
      readonly uniqueCode: CodecTypes['pg/uuid@1']['input'];
      readonly checkedInAt: CodecTypes['pg/timestamptz-string@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly city: CodecTypes['pg/text@1']['input'];
      readonly state:
        | 'AC'
        | 'AL'
        | 'AP'
        | 'AM'
        | 'BA'
        | 'CE'
        | 'DF'
        | 'ES'
        | 'GO'
        | 'MA'
        | 'MT'
        | 'MS'
        | 'MG'
        | 'PA'
        | 'PB'
        | 'PR'
        | 'PE'
        | 'PI'
        | 'RJ'
        | 'RN'
        | 'RS'
        | 'RO'
        | 'RR'
        | 'SC'
        | 'SP'
        | 'SE'
        | 'TO';
      readonly cpf: CodecTypes['pg/text@1']['input'];
      readonly tier: 'free' | 'verified' | 'pro';
      readonly passwordHash: CodecTypes['pg/text@1']['input'];
      readonly birthDate: CodecTypes['pg/date-temporal@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly events: {
      readonly capacity: CodecTypes['pg/int4@1']['output'];
      readonly city: CodecTypes['pg/text@1']['output'];
      readonly created_at: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly creator_id: CodecTypes['pg/uuid@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly location: CodecTypes['pg/text@1']['output'];
      readonly min_age: CodecTypes['pg/int4@1']['output'];
      readonly sold_tickets: CodecTypes['pg/int4@1']['output'];
      readonly starts_at: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly state:
        | 'AC'
        | 'AL'
        | 'AP'
        | 'AM'
        | 'BA'
        | 'CE'
        | 'DF'
        | 'ES'
        | 'GO'
        | 'MA'
        | 'MT'
        | 'MS'
        | 'MG'
        | 'PA'
        | 'PB'
        | 'PR'
        | 'PE'
        | 'PI'
        | 'RJ'
        | 'RN'
        | 'RS'
        | 'RO'
        | 'RR'
        | 'SC'
        | 'SP'
        | 'SE'
        | 'TO';
      readonly status: 'draft' | 'published' | 'finished' | 'canceled';
      readonly ticket_price_cents: CodecTypes['pg/int4@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly updated_at: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly orders: {
      readonly created_at: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly event_id: CodecTypes['pg/uuid@1']['output'];
      readonly expires_at: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly payment_id: CodecTypes['pg/text@1']['output'] | null;
      readonly quantity: CodecTypes['pg/int4@1']['output'];
      readonly status: 'pending' | 'paid' | 'expired' | 'canceled';
      readonly unit_price_cents: CodecTypes['pg/int4@1']['output'];
      readonly updated_at: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly user_id: CodecTypes['pg/uuid@1']['output'];
    };
    readonly tickets: {
      readonly checked_in_at: CodecTypes['pg/timestamptz-string@1']['output'] | null;
      readonly created_at: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly event_id: CodecTypes['pg/uuid@1']['output'];
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly order_id: CodecTypes['pg/uuid@1']['output'];
      readonly unique_code: CodecTypes['pg/uuid@1']['output'];
      readonly updated_at: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly user_id: CodecTypes['pg/uuid@1']['output'];
    };
    readonly users: {
      readonly birth_date: CodecTypes['pg/date-temporal@1']['output'];
      readonly city: CodecTypes['pg/text@1']['output'];
      readonly cpf: CodecTypes['pg/text@1']['output'];
      readonly created_at: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly password_hash: CodecTypes['pg/text@1']['output'];
      readonly state:
        | 'AC'
        | 'AL'
        | 'AP'
        | 'AM'
        | 'BA'
        | 'CE'
        | 'DF'
        | 'ES'
        | 'GO'
        | 'MA'
        | 'MT'
        | 'MS'
        | 'MG'
        | 'PA'
        | 'PB'
        | 'PR'
        | 'PE'
        | 'PI'
        | 'RJ'
        | 'RN'
        | 'RS'
        | 'RO'
        | 'RR'
        | 'SC'
        | 'SP'
        | 'SE'
        | 'TO';
      readonly tier: 'free' | 'verified' | 'pro';
      readonly updated_at: CodecTypes['pg/timestamptz-string@1']['output'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;